Add the first draft of the new selection menu
