+++ common/components/settings/src/commonMain/kotlin/io/stardewvalleydesigner/settings/wordlists/WordList.kt
@@ -16,7 +16,6 @@
 
 package io.stardewvalleydesigner.settings.wordlists
 
-import io.stardewvalleydesigner.component.editor.menus.EntitySelectionRoot
 import io.stardewvalleydesigner.component.editor.modules.toolkit.ShapeType
 import io.stardewvalleydesigner.component.editor.modules.toolkit.ToolType
 import io.stardewvalleydesigner.designformat.models.OptionsItemValue
@@ -165,11 +164,6 @@ sealed interface WordList {
 
 
     // Editor Screen
-
-    /**
-     * Get menu title by its root.
-     */
-    fun menuTitle(root: EntitySelectionRoot): String
 
     fun optionTitle(option: OptionsItemValue): String
