+++ common/ui-components/editor-menus/save-design-img/src/commonMain/kotlin/io/stardewvalleydesigner/ui/component/editor/screen/bottommenu/SaveDesignAsImageButton.kt
@@ -33,6 +33,7 @@ import io.stardewvalleydesigner.ui.component.editor.res.ImageResources
 import io.stardewvalleydesigner.ui.component.editor.res.ImageResourcesProvider
 import io.stardewvalleydesigner.ui.component.editor.screen.bottommenu.savedesignasimg.DesignRenderer
 import io.stardewvalleydesigner.ui.component.editor.screen.bottommenu.savedesignasimg.nowFormatted
+import io.stardewvalleydesigner.ui.component.editor.screen.topmenu.BottomMenuIconButton
 import io.stardewvalleydesigner.ui.component.settings.GlobalSettings
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
