+++ common/ui-components/editor-menus/save-design/src/jvmMain/kotlin/io/stardewvalleydesigner/ui/component/editor/screen/bottommenu/SaveDesignButton.kt
@@ -27,6 +27,7 @@ import io.stardewvalleydesigner.designformat.models.Options
 import io.stardewvalleydesigner.designformat.models.Palette
 import io.stardewvalleydesigner.kmplib.dispatcher.PlatformDispatcher
 import io.stardewvalleydesigner.ui.component.editor.screen.bottommenu.savedesign.DesignSaver
+import io.stardewvalleydesigner.ui.component.editor.screen.topmenu.BottomMenuIconButton
 import io.stardewvalleydesigner.ui.component.settings.GlobalSettings
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
